requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-dotenv
alembic
passlib[argon2]
python-jose[cryptography]
pytest
httpx
scikit-learn
pandas
joblib
openai
